Swap disable GameInput to Toggle

+++ Hellbomb Script.ps1
@@ -394,21 +394,27 @@ Function Install-VCRedist {
 }
 Function Disable-GameInput {
     Write-Host "Disabling GameInput..." -ForegroundColor Cyan
-    try { 
-        $gameInputSvc = Get-Service -Name "GameInputSvc"
-        If($gameInputSvc.StartType -eq "Disabled") {
-            Write-Host "GameInput is already disabled." -ForegroundColor Green
-            return
+    Try { 
+        	$gameInputSvc = Get-Service -Name "GameInputSvc"
+		}
+		Catch { 
+        	Write-Host "GameInput is not installed." -ForegroundColor Green
+       	 	Break
+    	}
+        If($gameInputSvc.StartType -eq "Enabled") {
+			Write-Host "Disabling GameInput..." -ForegroundColor Cyan
+			Stop-Service "GameInputSvc"
+	    	Set-Service "GameInputSvc" -StartupType Disabled
+	    	Write-Host "GameInput now Disabled." -ForegroundColor Green
+			Break
+		If($gameInputSvc.StartType -eq "Disabled") {
+            Write-Host "Enabling GameInput..." -ForegroundColor Cyan
+			Set-Service "GameInputSvc" -StartupType Disabled
+			Stop-Service "GameInputSvc"
+			Write-Host "GameInput now Enabled." -ForegroundColor Yellow
+            Break
         }
-    }
-    Catch { 
-        Write-Host "GameInput is not installed." -ForegroundColor Yellow
-        return
-    }
-
-    Stop-Service "GameInputSvc"
-    Set-Service "GameInputSvc" -StartupType Disabled
-    Write-Host "GameInput Disabled." -ForegroundColor Green
+    
 }
 Function Find-BlacklistedDrivers {
     $BadDeviceList = @('A-Volute', 'Hamachi', 'Nahimic', 'LogMeIn Hamachi', 'Sonic')
@@ -1772,7 +1778,7 @@ $Title = @(
         [ChoiceDescription]::new("🧹 Clear &Z Hostability Key$([Environment]::NewLine)", 'Fixes some game join issues by removing the current hostability key in user_settings.config'),
         [ChoiceDescription]::new("🔁 Re-install &GameGuard$([Environment]::NewLine)", 'Performs a full GameGuard re-install. If Windows Ransomware Protection is enabled, may trigger security alert.'),
         [ChoiceDescription]::new("🔁 Re&set Steam$([Environment]::NewLine)", 'Performs a reset of Steam. This can fix various issues including VRAM memory leaks.'),
-        [ChoiceDescription]::new("🗑️ &Disable GameInput$([Environment]::NewLine)", 'Disables Microsoft GameInput service.'),
+        [ChoiceDescription]::new("🗑️ &Disable GameInput$([Environment]::NewLine)", 'Toggles Microsoft GameInput service. May resolve some stuttering issues.'),
         [ChoiceDescription]::new("🗑️ &Uninstall VC++ Redists$([Environment]::NewLine)", 'Preps for installing VC++ Redists. Restart required.'),
         [ChoiceDescription]::new("➕ &Install VC++ Redists$([Environment]::NewLine)", 'Installs Microsoft Visual C++ Redistributables required by HD2. Fixes startup issues. Restart required.'),
         [ChoiceDescription]::new("🛠️ Set HD2 G&PU$([Environment]::NewLine)", 'Brings up the Windows GPU settings.'),
@@ -2020,3 +2026,4 @@ Get-IsProcessRunning $HelldiversProcess
 $script:InstalledProgramsList = Get-InstalledPrograms
 Write-Host "Building menu... $([Environment]::NewLine)$([Environment]::NewLine)"
 Menu
+
